+++ src/App/frontend/src/codegen/run.ts
@@ -4,7 +4,6 @@ import path from 'node:path';
 import { CodeGeneratorContext } from 'src/codegen/CodeGeneratorContext';
 import { generateAllCommonTypes, generateCommonTypeScript } from 'src/codegen/Common';
 import { LayoutSchemaV1 } from 'src/codegen/schemas/layout.schema.v1';
-import { LayoutSetsSchemaV1 } from 'src/codegen/schemas/layout-sets.schema.v1';
 import { LayoutSettingsSchemaV1 } from 'src/codegen/schemas/layoutSettings.schema.v1';
 import { saveFile, saveTsFile } from 'src/codegen/tools';
 import type { ComponentConfig } from 'src/codegen/ComponentConfig';
@@ -99,11 +98,7 @@ async function getComponentList(): Promise<[ComponentList, string[]]> {
   }
 
   const schemaProps: SchemaFileProps = { configMap, componentList, sortedKeys };
-  const schemas = [
-    new LayoutSchemaV1(schemaProps),
-    new LayoutSetsSchemaV1(schemaProps),
-    new LayoutSettingsSchemaV1(schemaProps),
-  ];
+  const schemas = [new LayoutSchemaV1(schemaProps), new LayoutSettingsSchemaV1(schemaProps)];
 
   const schemaPathBase = 'schemas/json/';
   for (const file of schemas) {