+++ src/Designer/frontend/libs/studio-pure-functions/src/ArrayUtils/ArrayUtils.ts
@@ -254,4 +254,8 @@ export class ArrayUtils {
   public static isArrayOfStrings(arg: unknown): arg is string[] {
     return Array.isArray(arg) && arg.every((item) => typeof item === 'string');
   }
+
+  public static filterNotNull<T>(array: (T | null)[]): T[] {
+    return array.filter((item): item is T => item !== null);
+  }
 }