+++ src/Designer/frontend/packages/shared/src/mocks/queriesMock.ts
@@ -40,6 +40,7 @@ import type { LayoutSets } from 'app-shared/types/api/LayoutSetsResponse';
 import type {
   IFrontEndSettings,
   ILayoutSettings,
+  IValidationOnNavigationLayoutSettings,
   ITextResourcesWithLanguage,
 } from 'app-shared/types/global';
 import type { WidgetSettingsResponse } from 'app-shared/types/widgetTypes';
@@ -128,6 +129,9 @@ export const queriesMock: ServicesContextProps = {
   getRepoDiff: jest.fn().mockImplementation(() => Promise.resolve<RepoDiffResponse>({})),
   getEnvironments: jest.fn().mockImplementation(() => Promise.resolve<Environment[]>([])),
   getFormLayoutSettings: jest.fn().mockImplementation(() => Promise.resolve<ILayoutSettings>({})),
+  getValidationOnNavigationLayoutSettings: jest
+    .fn()
+    .mockImplementation(() => Promise.resolve<IValidationOnNavigationLayoutSettings[]>([])),
   getFormLayouts: jest.fn().mockImplementation(() => Promise.resolve<FormLayoutsResponse>({})),
   getFormLayoutsV3: jest.fn().mockImplementation(() => Promise.resolve<FormLayoutsResponseV3>({})),
   getFrontEndSettings: jest.fn().mockImplementation(() => Promise.resolve<IFrontEndSettings>({})),
@@ -137,6 +141,9 @@ export const queriesMock: ServicesContextProps = {
   getLayoutSetsExtended: jest
     .fn()
     .mockImplementation(() => Promise.resolve<LayoutSetModel[]>(layoutSetsExtendedMock)),
+  getValidationOnNavigationLayoutSets: jest
+    .fn()
+    .mockImplementation(() => Promise.resolve({ show: [], page: '' })),
   getOptionListIds: jest.fn().mockImplementation(() => Promise.resolve<string[]>([])),
   getOptionList: jest.fn().mockImplementation(() => Promise.resolve<OptionList>([])),
   getOptionLists: jest.fn().mockImplementation(() => Promise.resolve<OptionListsResponse>([])),
@@ -285,6 +292,7 @@ export const queriesMock: ServicesContextProps = {
   deleteLayoutSet: jest.fn().mockImplementation(() => Promise.resolve()),
   deleteOptionList: jest.fn().mockImplementation(() => Promise.resolve()),
   deleteOrgCodeList: jest.fn().mockImplementation(() => Promise.resolve()),
+  deleteValidationOnNavigationLayoutSets: jest.fn().mockImplementation(() => Promise.resolve()),
   discardChanges: jest.fn().mockImplementation(() => Promise.resolve()),
   generateModels: jest.fn().mockImplementation(() => Promise.resolve()),
   importCodeListFromOrgToApp: jest.fn().mockImplementation(() => Promise.resolve<OptionList>([])),
@@ -309,6 +317,7 @@ export const queriesMock: ServicesContextProps = {
   updateAppPolicy: jest.fn().mockImplementation(() => Promise.resolve()),
   updateAppMetadata: jest.fn().mockImplementation(() => Promise.resolve()),
   updateAppConfig: jest.fn().mockImplementation(() => Promise.resolve()),
+  updateValidationOnNavigationLayoutSets: jest.fn().mockImplementation(() => Promise.resolve()),
   updateOptionList: jest.fn().mockImplementation(() => Promise.resolve()),
   updateOptionListId: jest.fn().mockImplementation(() => Promise.resolve()),
   updateOrgCodeListId: jest.fn().mockImplementation(() => Promise.resolve()),