+++ src/Designer/frontend/packages/ux-editor/src/components/Settings/SettingsNavigation/ValidateNavigation/TargetVariants/ValidateSelectedTasksConfig.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import { useTranslation } from 'react-i18next';
 import { ValidateNavigationConfig } from '../ValidateNavigationConfig';
 import {
   Scope,
@@ -11,8 +10,6 @@ import type { ExternalConfigWithId, InternalConfigState } from '../utils/Validat
 import { useConvertToInternalConfig } from '../utils/useConvertToInternalConfig';
 
 export const ValidateSelectedTasksConfig = () => {
-  const { t } = useTranslation();
-
   const [tempExtConfigs, setTempExtConfigs] = useState<ExternalConfigWithId[]>( // This is just to simulate the save functionality, in real implementation this would be handled differently
     withUniqueIds(dummyDataTasks),
   );
@@ -42,16 +39,16 @@ export const ValidateSelectedTasksConfig = () => {
         internalConfigs.map((conf, index) => (
           <ValidateNavigationConfig
             key={conf.id}
-            propertyLabel={t('ux_editor.settings.navigation_validation_button_label')}
             scope={Scope.SelectedTasks}
             config={conf}
+            existingConfigs={internalConfigs}
             onSave={(newConf) => handleSave(newConf, conf.id)}
             onDelete={() => handleDelete(conf.id)}
           />
         ))}
       <ValidateNavigationConfig
-        propertyLabel={t('ux_editor.settings.navigation_validation_button_label')}
         scope={Scope.SelectedTasks}
+        existingConfigs={internalConfigs}
         onSave={(newConf) => handleSave(newConf)}
       />
     </>