+++ src/Designer/frontend/packages/ux-editor/src/components/Settings/SettingsNavigation/ValidateNavigation/ValidateCardContent/ValidateCardContent.test.tsx
@@ -86,7 +86,7 @@ const renderValidateCardContent = ({
   };
 
   renderWithProviders(
-    <ValidateCardContent scope={scope} config={mockConfig} onChange={onChange} />,
+    <ValidateCardContent scope={scope} newConfig={mockConfig} onChange={onChange} />,
     { queryClient },
   );
 };