+++ src/Designer/frontend/packages/ux-editor/src/components/Settings/SettingsNavigation/ValidateNavigation/ValidateNavigationConfig.tsx
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
-import { StudioConfigCard, StudioProperty } from '@studio/components';
+import { StudioAlert, StudioLabel, StudioConfigCard, StudioProperty } from '@studio/components';
 import {
   type Scope,
+  isRuleDuplicateInScope,
   getCardLabel,
   getDefaultConfig,
   getValuesToDisplay,
@@ -11,31 +12,38 @@ import { useTranslation } from 'react-i18next';
 import classes from './ValidateNavigationConfig.module.css';
 import { ValidateCardContent } from './ValidateCardContent/ValidateCardContent';
 import type { InternalConfigState } from './utils/ValidateNavigationTypes';
+import cn from 'classnames';
 
 export type ValidateNavigationConfigProps = {
-  propertyLabel: string;
   scope: Scope;
   config?: InternalConfigState;
+  existingConfigs?: InternalConfigState[];
   onSave: (config: InternalConfigState) => void;
   onDelete?: () => void;
 };
 
 export const ValidateNavigationConfig = ({
-  propertyLabel,
   scope,
   config,
+  existingConfigs,
   onSave,
   onDelete,
 }: ValidateNavigationConfigProps) => {
   const [isEditMode, setIsEditMode] = useState(false);
+  const { t } = useTranslation();
+
+  const getButtonLabel = (currentConfig: InternalConfigState) => {
+    return !currentConfig && t('ux_editor.settings.navigation_validation_button_rule_undefined');
+  };
 
   if (!isEditMode) {
     return (
       <StudioProperty.Button
         onClick={() => setIsEditMode(true)}
-        property={propertyLabel}
+        property={getButtonLabel(config)}
+        title={config && t('ux_editor.settings.navigation_validation_button_rule_defined')}
         value={config && <DisplayValues {...config} />}
-        className={classes.configWrapper}
+        className={cn(classes.configWrapper, { [classes.configDefined]: config })}
       />
     );
   }
@@ -44,6 +52,7 @@ export const ValidateNavigationConfig = ({
     <ValidateCard
       scope={scope}
       config={config}
+      existingConfigs={existingConfigs}
       setIsEditMode={setIsEditMode}
       onSave={onSave}
       onDelete={onDelete}
@@ -54,20 +63,34 @@ export const ValidateNavigationConfig = ({
 type ValidateCardProps = {
   scope: Scope;
   config?: InternalConfigState;
+  existingConfigs?: InternalConfigState[];
   setIsEditMode: (isEditMode: boolean) => void;
   onSave: (config: InternalConfigState) => void;
   onDelete?: () => void;
 };
 
-const ValidateCard = ({ scope, config, setIsEditMode, onSave, onDelete }: ValidateCardProps) => {
+const ValidateCard = ({
+  scope,
+  config,
+  existingConfigs,
+  setIsEditMode,
+  onSave,
+  onDelete,
+}: ValidateCardProps) => {
   const { t } = useTranslation();
-  const [currentConfig, setCurrentConfig] = useState<InternalConfigState>(
+  const [newConfig, setNewConfig] = useState<InternalConfigState>(
     config || getDefaultConfig(scope),
   );
-  const isFormValid = validateForm({ scope, config, currentConfig });
+  const isFormValid = validateForm({ scope, config, newConfig });
+  const isRuleDuplicate = isRuleDuplicateInScope({
+    scope,
+    newConfig,
+    existingConfigs,
+    isFormValid,
+  });
 
   const update = (updates: Partial<InternalConfigState>) => {
-    setCurrentConfig((prev) => ({ ...prev, ...updates }));
+    setNewConfig((prev) => ({ ...prev, ...updates }));
   };
 
   const handleDelete = () => {
@@ -76,7 +99,7 @@ const ValidateCard = ({ scope, config, setIsEditMode, onSave, onDelete }: Valida
   };
 
   const handleSaveAndClose = () => {
-    onSave(currentConfig);
+    onSave(newConfig);
     setIsEditMode(false);
   };
 
@@ -93,7 +116,12 @@ const ValidateCard = ({ scope, config, setIsEditMode, onSave, onDelete }: Valida
         isDeleteDisabled={!config}
       />
       <StudioConfigCard.Body>
-        <ValidateCardContent scope={scope} config={currentConfig} onChange={update} />
+        <ValidateCardContent scope={scope} newConfig={newConfig} onChange={update} />
+        {isRuleDuplicate && (
+          <StudioAlert data-color='info'>
+            {t('ux_editor.settings.navigation_validation_alert_message')}
+          </StudioAlert>
+        )}
       </StudioConfigCard.Body>
       <StudioConfigCard.Footer
         saveLabel={t('general.save')}
@@ -108,11 +136,17 @@ const ValidateCard = ({ scope, config, setIsEditMode, onSave, onDelete }: Valida
 
 const DisplayValues = (config: InternalConfigState) => {
   const valueToDisplay = getValuesToDisplay(config);
+  const { t } = useTranslation();
+  const translateKeyToDisplay = (key: string) => {
+    return t(`ux_editor.settings.navigation_validation_view_mode_label_${key}`);
+  };
 
   return (
     <div>
       {Object.entries(valueToDisplay).map(([key, value]) => (
-        <div key={key}> {value}</div>
+        <div key={key}>
+          <StudioLabel>{translateKeyToDisplay(key)}:</StudioLabel> {value}
+        </div>
       ))}
     </div>
   );